ref: split taurus codebase into core, taurus (service), tests (test-suite)
