+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["taurus"]
+members = [ "crates/core", "crates/taurus", "crates/tests" ]
 resolver = "3"
 
 [workspace.package]
@@ -19,3 +19,6 @@ async-nats = "0.46.0"
 prost = "0.14.1"
 tonic-health = "0.14.1"
 tonic = "0.14.1"
+
+[workspace.dependencies.core]
+path = "./crates/core"
