+++ crates/core/src/context/context.rs
@@ -1,6 +1,7 @@
-use crate::error::RuntimeError;
 use std::collections::HashMap;
 use tucana::shared::{InputType, ReferenceValue, Value, value::Kind};
+
+use crate::runtime::error::RuntimeError;
 
 #[derive(Clone)]
 pub enum ContextResult {
