+++ crates/core/src/context/macros.rs
@@ -6,7 +6,7 @@ macro_rules! args {
         let __expected: usize = 0usize $(+ { let _ = ::core::any::type_name::<$ty>(); 1usize })*;
         if $args_ident.len() != __expected {
             return $crate::context::signal::Signal::Failure(
-                $crate::error::RuntimeError::simple(
+                $crate::runtime::error::RuntimeError::simple(
                     "InvalidArgumentRuntimeError",
                     format!("Expected {__expected} args but received {}", $args_ident.len()),
                 )
@@ -39,7 +39,7 @@ macro_rules! args {
 macro_rules! no_args {
     ($args_ident:ident) => {
         if !$args_ident.is_empty() {
-            return $crate::context::signal::Signal::Failure($crate::error::RuntimeError::simple(
+            return $crate::context::signal::Signal::Failure($crate::runtime::error::RuntimeError::simple(
                 "InvalidArgumentRuntimeError",
                 format!("Expected 0 args but received {}", $args_ident.len()),
             ));
