+++ crates/core/src/context/registry.rs
