+++ crates/core/src/debug/render.rs
