+++ crates/core/src/runtime/functions/array.rs
@@ -5,10 +5,11 @@ use tucana::shared::{ListValue, Value, value::Kind};
 
 use crate::context::argument::Argument;
 use crate::context::argument::ParameterNode::{Eager, Lazy};
+use crate::context::context::Context;
 use crate::context::macros::args;
 use crate::context::registry::{HandlerFn, HandlerFunctionEntry, IntoFunctionEntry};
 use crate::context::signal::Signal;
-use crate::{context::context::Context, error::RuntimeError};
+use crate::runtime::error::RuntimeError;
 
 pub fn collect_array_functions() -> Vec<(&'static str, HandlerFunctionEntry)> {
     vec![
