+++ crates/core/src/runtime/functions/boolean.rs
@@ -1,8 +1,9 @@
 use crate::context::argument::Argument;
+use crate::context::context::Context;
 use crate::context::macros::args;
 use crate::context::registry::{HandlerFn, HandlerFunctionEntry, IntoFunctionEntry};
 use crate::context::signal::Signal;
-use crate::{context::context::Context, error::RuntimeError};
+use crate::runtime::error::RuntimeError;
 use tucana::shared::{Value, value::Kind};
 
 pub fn collect_boolean_functions() -> Vec<(&'static str, HandlerFunctionEntry)> {
