+++ crates/core/src/runtime/functions/number.rs
@@ -6,7 +6,8 @@ use crate::context::argument::Argument;
 use crate::context::macros::{args, no_args};
 use crate::context::registry::{HandlerFn, HandlerFunctionEntry, IntoFunctionEntry};
 use crate::context::signal::Signal;
-use crate::{context::context::Context, error::RuntimeError};
+use crate::runtime::error::RuntimeError;
+use crate::{context::context::Context};
 
 pub fn collect_number_functions() -> Vec<(&'static str, HandlerFunctionEntry)> {
     vec![
