+++ crates/core/src/runtime/functions/text.rs
@@ -2,7 +2,8 @@ use crate::context::argument::Argument;
 use crate::context::macros::args;
 use crate::context::registry::{HandlerFn, HandlerFunctionEntry, IntoFunctionEntry};
 use crate::context::signal::Signal;
-use crate::{context::context::Context, error::RuntimeError};
+use crate::runtime::error::RuntimeError;
+use crate::{context::context::Context};
 use base64::Engine;
 use tucana::shared::{ListValue, Value, value::Kind};
