+++ crates/taurus/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "taurus"
-version = "0.1.0"
-edition = "2024"
+version.workspace = true
+edition.workspace = true
 
 [dependencies]
 code0-flow = { workspace = true, features = ["flow_service"] }
@@ -16,3 +16,4 @@ async-nats = { workspace = true }
 prost = { workspace = true }
 tonic-health = { workspace = true }
 tonic = { workspace = true }
+core = { workspace = true }
