+++ crates/tests/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "tests"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
